chore: start test setup for app
