+++ deployables/app/app/routes/edit-route.spec.ts
@@ -1,0 +1,5 @@
+import { describe, it } from 'vitest'
+
+describe('Edit route', () => {
+  it('shows the name of a route', () => {})
+})
