+++ deployables/app/package.json
@@ -6,7 +6,8 @@
     "build": "react-router build",
     "dev": "react-router dev",
     "start": "wrangler dev",
-    "check-types": "react-router typegen && tsc -b"
+    "check-types": "react-router typegen && tsc -b",
+    "test": "vitest"
   },
   "dependencies": {
     "@react-router/node": "^7.1.1",
@@ -21,15 +22,20 @@
     "@cloudflare/workers-types": "4.20241230.0",
     "@hiogawa/vite-node-miniflare": "0.1.1",
     "@react-router/dev": "^7.1.1",
+    "@testing-library/jest-dom": "^6.4.6",
+    "@testing-library/react": "^16.0.1",
+    "@testing-library/user-event": "14.5.2",
     "@types/node": "^22.0.0",
     "@types/react": "^19.0.1",
     "@types/react-dom": "^19.0.1",
     "autoprefixer": "^10.4.20",
+    "eslint": "^9.7.0",
     "postcss": "^8.4.49",
     "tailwindcss": "^3.4.16",
     "typescript": "^5.7.2",
     "vite": "^5.4.11",
     "vite-tsconfig-paths": "^5.1.4",
+    "vitest": "2.1.8",
     "wrangler": "^3.87.0"
   }
 }
