+++ deployables/app/test-utils/index.ts
@@ -1,0 +1,1 @@
+export { render } from './render'
