+++ deployables/app/vitest.setup.mts
@@ -1,0 +1,1 @@
+import '@testing-library/jest-dom/vitest'
